refactor: isolate types and code related to generation
